+++ package.json
@@ -11,9 +11,10 @@
   },
   "scripts": {
     "build": "wp-scripts build",
-    "format:js": "wp-scripts format",
+    "format": "wp-scripts format",
     "lint:css": "wp-scripts lint-style",
     "lint:js": "wp-scripts lint-js",
-    "start": "wp-scripts start"
+    "start": "wp-scripts start",
+    "packages-update": "wp-scripts packages-update"
   }
 }
